Add function to determine if a grade has expired
