+++ server/src/controllers/utils/date.ts
@@ -4,9 +4,6 @@
 
 import { DateOnlyString } from 'aalto-grades-common/types';
 
-import Attainment from '../../database/models/attainment';
-import AttainmentGrade from '../../database/models/attainmentGrade';
-
 export function toDateOnlyString(date: Date | string): DateOnlyString {
   const dateOnlyRegExp: RegExp =
     /^\d{4}[/-](0?[1-9]|1[012])[/-](0?[1-9]|[12][0-9]|3[01])$/;
@@ -23,43 +20,4 @@ export function toDateOnlyString(date: Date | string): DateOnlyString {
     return validate(date.toISOString().split('T')[0]);
 
   return toDateOnlyString(new Date(date));
-}
-
-export async function getDateOfLatestGrade(
-  userId: number, assessmentModelId: number
-): Promise<Date> {
-  const grades: Array<AttainmentGrade> = await AttainmentGrade.findAll({
-    where: {
-      userId: userId,
-      manual: true
-    },
-    include: [
-      {
-        model: Attainment,
-        where: {
-          assessmentModelId: assessmentModelId
-        }
-      }
-    ]
-  });
-
-  const dates: Array<Date> = grades.map(
-    (grade: AttainmentGrade) => new Date(grade.date)
-  );
-
-  let maxSoFar: Date | null = null;
-  for (const date of dates) {
-    if (!maxSoFar || date > maxSoFar) {
-      maxSoFar = date;
-    }
-  }
-
-  if (maxSoFar) {
-    return maxSoFar;
-  } else {
-    throw new Error(
-      `failed to find the date of the latest grade, user ${userId} has`
-      + ` no grades for assessment model ${assessmentModelId}`
-    )
-  }
 }
